+++ src/export.h
@@ -1,2 +1,4 @@
-extern "C" double _nextbyte(double x, int that);
-extern "C" double nextbyte(double x, int that);
+#include <inttypes.h>
+
+extern "C" double _nextbyte(double x, uint64_t that);
+extern "C" double nextbyte(double x, uint64_t that);
